refactor: update Makefile and header files for improved build configuration and memory management

+++ src/isr80h/misc.c
@@ -4,7 +4,7 @@
 
 void *isr80h_comman0_sum(struct interrupt_frame *frame)
 {
-	int v2 = (int)task_get_stack_item(task_current(), 1);
-	int v1 = (int)task_get_stack_item(task_current(), 0);
+	intptr_t v2 = (uintptr_t)task_get_stack_item(task_current(), 1);
+	intptr_t v1 = (uintptr_t)task_get_stack_item(task_current(), 0);
 	return (void *)(v1 + v2);
 }
